F-04 Add .env variables
